requirements.txt
torch
torchvision
timm
flask
Pillow
gunicorn
